% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_data}
\alias{age_data}
\title{Age and Demographics data}
\format{
\subsection{\code{age_data}}{

A list of 3 items.
\describe{
\item{xage}{data frame of 8856 observations on 13 variables}
\item{yage}{Randomly generated age in the range above}
\item{gbage}{gradient boosting model for median age given x}
}
}
}
\source{
The Elements of Statistical Learning, Data Mining, Second Edition, by Hastie,
Tibshirani, and Friedman.
}
\usage{
age_data
}
\description{
The data come from 9243 questionnaires filled out by shopping mall
customers in the San Francisco Bay Area (Impact Resources, Inc.,
Columbus, OH). Here we attempt to estimate a persons age as a
function of the other 13 demographic variables. For this data set
age value is reported as being in one of seven intervals \verb{\{13-17, 18-24, 25-34, 35-44, 45-54, 55-64, >= 65\}}. Each persons age is
randomly generated uniformly within its corresponding reported
interval. For the last interval an exponential distribution was
used with mean corresponding to life expectancy after reaching age
65.
}
\keyword{datasets}
