% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contree-package.R
\docType{package}
\name{conTree-package}
\alias{conTree-package}
\title{Contrast and Boosted Trees}
\description{
Contrast trees represent a new approach for assessing the accuracy
of many types of machine learning estimates that are not amenable
to standard (cross) validation methods. In situations where
inaccuracies are detected, boosted contrast trees can often improve
performance. Functions are provided to to build such trees in
addition to a special case, distribution boosting, an assumption
free method for estimating the full probability distribution of an
outcome variable given any set of joint input predictor variable
values.
}
\references{
Jerome Friedman (2019). \emph{Contrast Trees and
Distribution Boosting} \url{https://arxiv.org/abs/1912.03785}
}
\author{
Original code (C) by Jerome H. Friedman, minor modifications,
formatting, and packaging by Balasubramanian Narasimhan
}
