% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{prune}
\alias{prune}
\title{Prune a contrast tree}
\usage{
prune(tree, thr = 0.1)
}
\arguments{
\item{tree}{a tree model object output from contrast}

\item{thr}{a split improvement threshold, default is 0.1}
}
\value{
a bottom-up pruned tree with splits corresponding to improvement less than threshold \code{thr} removed
}
\description{
Prune a contrast tree
}
