% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_rfun.R
\name{save_rfun}
\alias{save_rfun}
\title{Save the function f for calling from fortran}
\usage{
save_rfun(f)
}
\arguments{
\item{f}{the R function to be called using \code{.Fortran}}
}
\value{
TRUE, invisibly.
}
\description{
Save the function f for calling from fortran
}
