% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{treesum}
\alias{treesum}
\title{Print terminal node x-region boundaries}
\usage{
treesum(tree, n = 0)
}
\arguments{
\item{tree}{model object output from contrast() or prune()}

\item{n}{the number of nodes displayed in order of criterion value (n=0 displays all terminal nodes)}
}
\description{
Print terminal node x-region boundaries
}
\details{
The predictor variable x-boundaries defining each terminal node are printed.

For numeric variables: variable | sign | value
\itemize{
\item sign + => value=lower boundary on variable
\item sign - => value upper boundary on variable
}

For categorical variables:  cat variable | sign | set of values
\itemize{
\item sign + => values in node
\item sign - => values not in node (compliment values in node) graphical representations of terminal node contrasts depend on the tree type
}
}
