% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{ydist}
\alias{ydist}
\title{Generate sample from estimated \eqn{p(y | x)} for type = 'dist' only}
\usage{
ydist(x, z, model, num = model$niter)
}
\arguments{
\item{x}{vector of predictor variable values for a (single) observation}

\item{z}{sample of z-values drawn from \eqn{p(z | x)}}

\item{model}{model object output from modtrast()}

\item{num}{number of trees used to compute model values}
}
\value{
vector of \code{length(z)} containing transformed values representing \eqn{p(y | x)}
}
\description{
Generate sample from estimated \eqn{p(y | x)} for type = 'dist' only
}
