% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{lofcurve}
\alias{lofcurve}
\title{Show graphical terminal node summaries}
\usage{
lofcurve(x, y, z, tree, w = rep(1, length(y)), doplot = TRUE, ylim = NULL)
}
\arguments{
\item{x}{training input predictor data matrix or data frame in same format as in contrast()}

\item{y}{vector, or matrix containing training data input outcome values or censoring intervals for each observation in same format as in contrast()}

\item{z}{vector containing values of a second contrasting quantity for each observation in same observation format as in contrast ()}

\item{tree}{model object output from contrast() or prune()}

\item{w}{observation weights}

\item{doplot}{a flag to display/not display graphical plots}

\item{ylim}{y-axis limit}
}
\value{
a named list of plotted \code{x} and \code{y} points
}
\description{
Show graphical terminal node summaries
}
