% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{predtrast}
\alias{predtrast}
\title{Predict y-values from boosted contrast model}
\usage{
predtrast(x, z, model, num = model$niter)
}
\arguments{
\item{x}{x-values for new data}

\item{z}{z-values for new data}

\item{model}{model object output from modtrast()}

\item{num}{number of trees used to compute model values}
}
\value{
predicted y-values for new data from model
}
\description{
Predict y-values from boosted contrast model
}
