% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{treesum}
\alias{treesum}
\title{Print terminal node x-region boundaries}
\usage{
treesum(tree, nodes = NULL)
}
\arguments{
\item{tree}{model object output from contrast() or prune()}

\item{nodes}{vector of terminal node identifiers for the tree specifying the desired regions. The default is all terminal nodes.}
}
\value{
No return value (invisble \code{NULL})
}
\description{
Print terminal node x-region boundaries
}
\details{
The predictor variable x-boundaries defining each terminal node are printed.

For numeric variables: variable | sign | value
\itemize{
\item sign + => value=lower boundary on variable
\item sign - => value upper boundary on variable
}

For categorical variables:  cat variable | sign | set of values
\itemize{
\item sign + => values in node
\item sign - => values not in node (compliment values in node) graphical representations of terminal node contrasts depend on the tree type
}
}
\seealso{
\code{\link[=contrast]{contrast()}}
}
