% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{xval}
\alias{xval}
\title{Cross-validate boosted contrast tree boosted with new data}
\usage{
xval(
  x,
  y,
  z,
  mdl,
  num = length(mdl$tree),
  del = 10,
  span = 0.15,
  ylab = "Average  Discrepancy",
  doplot = TRUE,
  doprint = TRUE
)
}
\arguments{
\item{x}{new data predictor variables is same format as input to modtrast}

\item{y}{new data y values is same format as input to modtrast}

\item{z}{new data z values is same format as input to modtrast}

\item{mdl}{model output from modtrast()}

\item{num}{number of trees used to compute model values}

\item{del}{plot discrepancy value computed every del-th iteration (tree)}

\item{span}{running median smoother span (\code{doplot=TRUE}, only)}

\item{ylab}{graphical parameter (\code{doplot=TRUE}, only)}

\item{doplot}{logical flag \code{TRUE/FALSE} implies do/don't show plot}

\item{doprint}{logical flag \code{TRUE/FALSE} implies do/don't print progress while executing}
}
\value{
a named list of two items: \code{ntree} the iteration numbers, and \code{error} the corresponding discrepancy values
}
\description{
Cross-validate boosted contrast tree boosted with new data
}
